requirements/development.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-bootstrap5==24.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-tables2==2.7.4
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
